Crate rule34

Structs

Client

A Rule34 Client

DeletedImageList

A list of deleted images

Note

NoteList

A list of notes for posts

NotesListQueryBuilder

A query builder to get notes.

Post

A Post

PostList

A list of posts

PostListQueryBuilder

A builder for post list api queries

SearchQueryBuilder

A helper to build a search query.

Tag

A Tag

TagKind

The tag kind

TagList

A list of tags

TagListQueryBuilder

A query builder to get tags

Url

A parsed URL record.

Enums

Error

Crate Error Type

PostStatus

A Post Status

Rating

A post rating

Constants

POST_LIST_LIMIT_MAX

The maximum number of responses per post list request

TAGS_LIST_LIMIT_MAX

The maximum number of responses per tags list request.