Struct rule34::Post

pub struct Post {

    pub height: NonZeroU64,
    pub score: i64,
    pub file_url: Url,
    pub parent_id: Option<NonZeroU64>,
    pub sample_url: Url,
    pub sample_width: NonZeroU64,
    pub sample_height: NonZeroU64,
    pub preview_url: Url,
    pub rating: Rating,
    pub tags: Box<str>,
    pub id: NonZeroU64,
    pub width: NonZeroU64,
    pub change: OffsetDateTime,
    pub md5: Md5Digest,
    pub creator_id: NonZeroU64,
    pub has_children: bool,
    pub created_at: OffsetDateTime,
    pub status: PostStatus,
    pub source: Option<Box<str>>,
    pub has_notes: bool,
    pub has_comments: bool,
    pub preview_width: NonZeroU64,
    pub preview_height: NonZeroU64,

}

A Post

Fields
height: NonZeroU64

The height of the original file.

score: i64

The number of up-votes minus the number of down-votes.

file_url: Url

The main file url

parent_id: Option<NonZeroU64>

The parent post id

sample_url: Url

The sample url

sample_width: NonZeroU64

The sample width

sample_height: NonZeroU64

The sample height

preview_url: Url

The preview url

rating: Rating

The image rating

tags: Box<str>

A list of tag names.

Tag names are separated by one or more spaces. There may or may not be a leading or trailing space. Tag names are always lowercase.

id: NonZeroU64

The id the post

width: NonZeroU64

image width

change: OffsetDateTime

The time of the last change.

This tracks at least the date of posting and tag edits. This is a unix timestamp.

md5: Md5Digest

The md5 hash of the file.

creator_id: NonZeroU64

The creator id.

has_children: bool

Whether this has children.

created_at: OffsetDateTime

The creation date of the post.

status: PostStatus

The status of the post.

source: Option<Box<str>>

The original source.

May or may not be a url, it is filled manually by users.

has_notes: bool

Whether the post has notes.

has_comments: bool

Whether this post has comments.

preview_width: NonZeroU64

The preview image width.

preview_height: NonZeroU64

The preview image height.

Implementations

impl Post

pub fn iter_tags(&self) -> impl Iterator<Item = &str>

Iter over the tags in this object.

There may be duplicate tags included.

pub fn get_html_post_url(&self) -> Url

Get the html post url for this post.

This allocates, so cache the result.